pyproject.toml

[project]
name = "vaulttool"
version = "0.1.1"
description = "Secure file encryption for secrets/configs"
authors = [
    { name = "Josef Fuchs" }
]
license = { text = "Apache-2" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer",
    "pyyaml"
]

[project.urls]
Homepage = "https://github.com/yourusername/vaulttool"

[project.scripts]
vaulttool = "vaulttool.__main__:main"

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = [
    "vaulttool/tests"
]

pythonpath = "."

[project.optional-dependencies]
dev = [
    "ruff",
    "pre-commit"
]
